+++ internal/querylog/querylog.go
@@ -25,18 +25,37 @@ type QueryLog interface {
 
 // Config - configuration object
 type Config struct {
-	Enabled           bool   // enable the module
-	FileEnabled       bool   // write logs to file
-	BaseDir           string // directory where log file is stored
-	Interval          uint32 // interval to rotate logs (in days)
-	MemSize           uint32 // number of entries kept in memory before they are flushed to disk
-	AnonymizeClientIP bool   // anonymize clients' IP addresses
-
-	// Called when the configuration is changed by HTTP request
+	// ConfigModified is called when the configuration is changed, for
+	// example by HTTP requests.
 	ConfigModified func()
 
-	// Register an HTTP handler
+	// HTTPRegister registers an HTTP handler.
 	HTTPRegister func(string, string, func(http.ResponseWriter, *http.Request))
+
+	// FindClient returns client information by their IDs.
+	FindClient func(ids []string) (c *Client, err error)
+
+	// BaseDir is the base directory for log files.
+	BaseDir string
+
+	// Interval is the interval for log rotation, in days.  After that
+	// period, the old log file will be renamed, NOT deleted, so the actual
+	// log retention time is twice the interval.
+	Interval uint32
+
+	// MemSize is the number of entries kept in a memory buffer before they
+	// are flushed to disk.
+	MemSize uint32
+
+	// Enabled tells if the query log is enabled.
+	Enabled bool
+
+	// FileEnabled tells if the query log writes logs to files.
+	FileEnabled bool
+
+	// AnonymizeClientIP tells if the query log should anonymize clients' IP
+	// addresses.
+	AnonymizeClientIP bool
 }
 
 // AddParams - parameters for Add()